ICMSLST-381 Refactor of the import Import Contact urls, views & templates.
