+++ web/domains/case/_import/forms.py
@@ -178,3 +178,45 @@ class EndorsementImportApplicationForm(forms.ModelForm):
     class Meta:
         model = models.EndorsementImportApplication
         fields = ("content",)
+
+
+class ImportContactPersonForm(forms.ModelForm):
+    last_name = forms.CharField(required=True, label="Surname")
+
+    class Meta:
+        model = models.ImportContact
+        fields = (
+            "first_name",
+            "last_name",
+            "street",
+            "city",
+            "postcode",
+            "region",
+            "country",
+            "dealer",
+        )
+        labels = {
+            "first_name": "First Name",
+            "last_name": "Surname",
+            "dealer": "Did you buy from a dealer?",
+        }
+
+
+class ImportContactLegalEntityForm(forms.ModelForm):
+    class Meta:
+        model = models.ImportContact
+        fields = (
+            "first_name",
+            "registration_number",
+            "street",
+            "city",
+            "postcode",
+            "region",
+            "country",
+            "dealer",
+        )
+        labels = {
+            "first_name": "Name of Legal Person",
+            "registration_number": "Registration Number",
+            "dealer": "Did you buy from a dealer?",
+        }
