+++ web/templates/partial/case/import/fa-oil/sidebar.html
@@ -6,7 +6,7 @@
     <a href="{{ icms_url('import:fa-oil:list-user-import-certificates', kwargs={'pk': process.pk}) }}">Certificates</a>
   </li>
   <li>
-    <a href="{{ icms_url('import:fa-oil:list-import-contacts', kwargs={'pk': process.pk}) }}">Details of Who Bought From</a>
+    <a href="{{ icms_url('import:fa-list-import-contacts', kwargs={'pk': process.pk}) }}">Details of Who Bought From</a>
   </li>
   <li>
     <a href="{{ icms_url('import:fa-oil:submit-oil', kwargs={'pk': process.pk}) }}">Submit</a>
