+++ web/templates/web/domains/case/import/fa-import-contacts/create.html
@@ -4,7 +4,8 @@
 
 
 {% block sidebar %}
-  {% include "partial/case/import/oil/sidebar.html" %}
+  {# TODO: Move to partials maybe? #}
+  {% include "web/domains/case/import/fa-import-contacts/sidebar.html" %}
 {% endblock %}
 
 {% block content_actions %}
@@ -12,7 +13,7 @@
   <ul class="menu-out flow-across">
     <li>
       <a
-        href="{{ icms_url('import:fa-oil:list-import-contacts', kwargs={'pk': process.pk}) }}"
+        href="{{ icms_url('import:fa-list-import-contacts', kwargs={'pk': process.pk}) }}"
         class="prev-link">
         Details of Who Bought From
       </a>
