+++ web/templates/web/domains/case/import/fa-import-contacts/edit.html
@@ -1,0 +1,1 @@
+{% extends "web/domains/case/import/fa-import-contacts/create.html" %}
