+++ web/templates/web/domains/case/import/fa-import-contacts/list.html
@@ -1,0 +1,36 @@
+{% extends "flow/task-base.html" %}
+
+
+{% block sidebar %}
+  {# TODO: Move to partials maybe? #}
+  {% include "web/domains/case/import/fa-import-contacts/sidebar.html" %}
+{% endblock %}
+
+{% block content_actions %}
+
+{% if process.process_type == "OpenIndividualLicenceApplication" %}
+  {% set url = 'import:fa-oil:edit-oil' %}
+  {% set label = 'Edit Open Individual Licence' %}
+{% else %}
+  {% set url = 'import:UNKNOWN-EDIT-ACTION' %}
+  {% set label = 'Unknown application type' %}
+{% endif %}
+
+<div class="content-actions">
+  <ul class="menu-out flow-across">
+    <li>
+      <a href="{{ icms_url(url, kwargs={'pk': process.pk}) }}" class="prev-link">
+        {{ label }}
+        {% if process.applicant_reference %}
+          '{{ process.applicant_reference }}'
+        {% endif %}
+      </a>
+    </li>
+  </ul>
+</div>
+{% endblock %}
+
+{% block main_content %}
+  <h3>Details of who bought from</h3>
+  {% include "partial/import-contacts/list.html" %}
+{% endblock %}
