+++ web/templates/web/domains/case/import/fa-oil/certificates/create.html
@@ -3,7 +3,7 @@
 {% import "forms/fields.html" as fields %}
 
 {% block sidebar %}
-  {% include "partial/case/import/oil/sidebar.html" %}
+  {% include "partial/case/import/fa-oil/sidebar.html" %}
 {% endblock %}
 
 {% block content_actions %}
