+++ web/templates/web/domains/case/import/fa-oil/certificates/list.html
@@ -2,7 +2,7 @@
 
 
 {% block sidebar %}
-  {% include "partial/case/import/oil/sidebar.html" %}
+  {% include "partial/case/import/fa-oil/sidebar.html" %}
 {% endblock %}
 
 {% block content_actions %}
