+++ web/templates/web/domains/case/import/fa-oil/certificates/view-verified.html
@@ -1,7 +1,7 @@
 {% extends "flow/task-base.html" %}
 
 {% block sidebar %}
-  {% include "partial/case/import/oil/sidebar.html" %}
+  {% include "partial/case/import/fa-oil/sidebar.html" %}
 {% endblock %}
 
 {% block content_actions %}
