--- web/templates/web/domains/case/import/fa-oil/import-contacts/edit.html
@@ -1,1 +1,0 @@
-{% extends "web/domains/case/import/fa-oil/import-contacts/create.html" %}
