--- web/templates/web/domains/case/import/fa-oil/import-contacts/list.html
@@ -1,28 +1,0 @@
-{% extends "flow/task-base.html" %}
-
-
-{% block sidebar %}
-  {% include "partial/case/import/oil/sidebar.html" %}
-{% endblock %}
-
-{% block content_actions %}
-<div class="content-actions">
-  <ul class="menu-out flow-across">
-    <li>
-      <a
-        href="{{ icms_url('import:fa-oil:edit-oil', kwargs={'pk': process.pk}) }}"
-        class="prev-link">
-        Edit Open Individual Licence
-        {% if process.applicant_reference %}
-          '{{ process.applicant_reference }}'
-        {% endif %}
-      </a>
-    </li>
-  </ul>
-</div>
-{% endblock %}
-
-{% block main_content %}
-  <h3>Details of who bought from</h3>
-  {% include "partial/import-contacts/list.html" %}
-{% endblock %}
