+++ web/templates/web/domains/case/import/fa-oil/submit.html
@@ -4,7 +4,7 @@
 
 
 {% block sidebar %}
-  {% include "partial/case/import/oil/sidebar.html" %}
+  {% include "partial/case/import/fa-oil/sidebar.html" %}
 {% endblock %}
